main page is rendered by Vue.
