+++ frontend/log/2017-09-05T10h_26m_34s
@@ -1,0 +1,50 @@
+2017/09/05 10:26:34 (main) The Logger has been initialized.
+2017/09/05 10:26:34 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 10:26:34 (main) The main page source code has been generated.
+2017/09/05 10:26:34 (main) Waiting request... 
+2017/09/05 10:26:36 (rootHandler) Main page access has been occurred.
+2017/09/05 10:26:36 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:26:38 (getDataFromDBByMajor) Request Major: 간호학과(야), Subjects Count: 15
+2017/09/05 10:26:38 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 10:26:41 (rootHandler) Main page access has been occurred.
+2017/09/05 10:26:41 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:44:10 (rootHandler) Main page access has been occurred.
+2017/09/05 10:44:11 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:44:15 (getDataFromDBByMajor) Request Major: 건축공학부, Subjects Count: 56
+2017/09/05 10:44:15 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 10:44:38 (rootHandler) Main page access has been occurred.
+2017/09/05 10:44:38 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:44:40 (getDataFromDBByMajor) Request Major: 건설환경공학과, Subjects Count: 39
+2017/09/05 10:44:40 (getDataFromDBByMajor) Request Major: 건설환경공학과, Subjects Count: 39
+2017/09/05 10:44:40 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 10:44:40 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 10:45:41 (rootHandler) Main page access has been occurred.
+2017/09/05 10:45:42 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:45:44 (getDataFromDBByMajor) Request Major: 간호학과, Subjects Count: 37
+2017/09/05 10:45:44 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 10:45:44 (getDataFromDBByMajor) Request Major: 간호학과, Subjects Count: 37
+2017/09/05 10:45:44 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 10:46:30 (rootHandler) Main page access has been occurred.
+2017/09/05 10:46:30 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:46:36 (getDataFromDBByMajor) Request Major: 건축공학부, Subjects Count: 56
+2017/09/05 10:46:36 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 10:46:36 (getDataFromDBByMajor) Request Major: 건축공학부, Subjects Count: 56
+2017/09/05 10:46:36 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 10:46:56 (rootHandler) Main page access has been occurred.
+2017/09/05 10:46:57 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:47:00 (getDataFromDBByMajor) Request Major: STS(과학기술학)전공, Subjects Count: 10
+2017/09/05 10:47:00 (getDataFromDBByMajor) Request Major: STS(과학기술학)전공, Subjects Count: 10
+2017/09/05 10:47:00 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 10:47:00 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 10:47:20 (rootHandler) Main page access has been occurred.
+2017/09/05 10:47:21 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:47:23 (getDataFromDBByMajor) Request Major: 건설환경공학과, Subjects Count: 39
+2017/09/05 10:47:23 (getDataFromDBByMajor) Request Major: 건설환경공학과, Subjects Count: 39
+2017/09/05 10:47:23 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 10:47:23 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 10:47:41 (rootHandler) Main page access has been occurred.
+2017/09/05 10:47:42 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 10:47:45 (getDataFromDBByMajor) Request Major: STS(과학기술학) 전공, Subjects Count: 10
+2017/09/05 10:47:45 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 10:47:45 (getDataFromDBByMajor) Request Major: STS(과학기술학) 전공, Subjects Count: 10
+2017/09/05 10:47:45 (sendSubjectTableHandler) The subject table html code has been sent to browser.
