+++ frontend/log/2017-09-05T11h_15m_39s
@@ -1,0 +1,9 @@
+2017/09/05 11:15:39 (main) The Logger has been initialized.
+2017/09/05 11:15:40 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:15:40 (main) The main page source code has been generated.
+2017/09/05 11:15:40 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:15:40 (main) The majorNameList has been generated.
+2017/09/05 11:15:40 (main) Waiting request... 
+2017/09/05 11:15:50 (rootHandler)  open ./webRoot/vue/getMajorNameList: no such file or directory
+2017/09/05 11:15:50 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 11:15:55 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
