+++ frontend/log/2017-09-05T11h_16m_16s
@@ -1,0 +1,7 @@
+2017/09/05 11:16:16 (main) The Logger has been initialized.
+2017/09/05 11:16:17 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:16:17 (main) The main page source code has been generated.
+2017/09/05 11:16:17 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:16:17 (main) The majorNameList has been generated.
+2017/09/05 11:16:17 (main) Waiting request... 
+2017/09/05 11:16:18 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
