+++ frontend/log/2017-09-05T11h_24m_47s
@@ -1,0 +1,10 @@
+2017/09/05 11:24:47 (main) The Logger has been initialized.
+2017/09/05 11:24:47 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:24:47 (main) The main page source code has been generated.
+2017/09/05 11:24:47 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:24:47 (main) The majorNameListJSON has been generated.
+2017/09/05 11:24:47 (main) Waiting request... 
+2017/09/05 11:24:52 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 11:24:53 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 11:26:13 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 11:26:14 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
