+++ frontend/log/2017-09-05T11h_45m_30s
@@ -1,0 +1,26 @@
+2017/09/05 11:45:30 (main) The Logger has been initialized.
+2017/09/05 11:45:30 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:45:30 (main) The main page source code has been generated.
+2017/09/05 11:45:31 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 11:45:31 (main) The majorNameListJSON has been generated.
+2017/09/05 11:45:31 (main) Waiting request... 
+2017/09/05 11:45:38 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 11:45:41 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 11:45:43 (getDataFromDBByMajor) Request Major: STS(과학기술학) 전공, Subjects Count: 10
+2017/09/05 11:45:43 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 11:45:43 (getDataFromDBByMajor) Request Major: STS(과학기술학) 전공, Subjects Count: 10
+2017/09/05 11:45:43 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 11:45:45 (rootHandler) Main page access has been occurred.
+2017/09/05 11:45:46 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 11:45:46 (rootHandler) Main page access has been occurred.
+2017/09/05 11:45:46 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 11:45:50 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 11:45:53 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 11:45:56 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 11:45:58 (getDataFromDBByMajor) Request Major: STS(과학기술학) 전공, Subjects Count: 10
+2017/09/05 11:45:58 (getDataFromDBByMajor) Request Major: STS(과학기술학) 전공, Subjects Count: 10
+2017/09/05 11:45:58 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 11:45:58 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 11:45:59 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 11:46:00 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 11:46:02 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
