+++ frontend/log/2017-09-05T12h_37m_20s
@@ -1,0 +1,10 @@
+2017/09/05 12:37:20 (main) The Logger has been initialized.
+2017/09/05 12:37:21 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 12:37:21 (main) The main page source code has been generated.
+2017/09/05 12:37:21 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 12:37:21 (main) The majorNameListJSON has been generated.
+2017/09/05 12:37:21 (main) Waiting request... 
+2017/09/05 12:37:23 (rootHandler) Main page access has been occurred.
+2017/09/05 12:37:23 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:37:26 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 12:37:26 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
