+++ frontend/log/2017-09-05T12h_40m_54s
@@ -1,0 +1,38 @@
+2017/09/05 12:40:54 (main) The Logger has been initialized.
+2017/09/05 12:40:54 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 12:40:54 (main) The main page source code has been generated.
+2017/09/05 12:40:55 (getMajorListFromDB) The subjects list has been generated.
+2017/09/05 12:40:55 (main) The majorNameListJSON has been generated.
+2017/09/05 12:40:55 (main) Waiting request... 
+2017/09/05 12:40:55 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 12:40:55 (rootHandler)  open ./webRoot/resources/js/moveToVue/subjectTable.js: no such file or directory
+2017/09/05 12:40:56 (rootHandler)  open ./webRoot/resources/js/moveToVue/subjectTable.js: no such file or directory
+2017/09/05 12:40:56 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:41:09 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 12:41:09 (rootHandler)  open ./webRoot/resources/js/moveToVue/subjectTable.js: no such file or directory
+2017/09/05 12:41:10 (rootHandler)  open ./webRoot/resources/js/moveToVue/subjectTable.js: no such file or directory
+2017/09/05 12:41:10 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:41:27 (rootHandler)  open ./webRoot/resources/js/moveToVue/subjectTable.js: no such file or directory
+2017/09/05 12:41:27 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:41:51 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 12:41:52 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:41:58 (getDataFromDBByMajor) Request Major: 건축공학부, Subjects Count: 56
+2017/09/05 12:41:58 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 12:41:58 (getDataFromDBByMajor) Request Major: 건축공학부, Subjects Count: 56
+2017/09/05 12:41:58 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 12:42:01 (rootHandler) Main page access has been occurred.
+2017/09/05 12:42:02 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:42:03 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 12:42:04 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:44:13 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 12:44:14 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:44:15 (getDataFromDBByMajor) Request Major: 간호학과(야), Subjects Count: 15
+2017/09/05 12:44:15 (getDataFromDBByMajor) Request Major: 간호학과(야), Subjects Count: 15
+2017/09/05 12:44:15 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 12:44:15 (sendSubjectTableHandler) The subject table html code has been sent to browser.
+2017/09/05 12:48:00 (rootHandler) Vue Main page access has been occurred.
+2017/09/05 12:48:01 (rootHandler)  open ./webRoot/favicon.ico: no such file or directory
+2017/09/05 12:48:02 (getDataFromDBByMajor) Request Major: 간호학과(야), Subjects Count: 15
+2017/09/05 12:48:02 (sendDataByMajorHandler) The subject data have been sent.
+2017/09/05 12:48:02 (getDataFromDBByMajor) Request Major: 간호학과(야), Subjects Count: 15
+2017/09/05 12:48:02 (sendSubjectTableHandler) The subject table html code has been sent to browser.
