+++ frontend/webRoot/resources/js/basket.js
@@ -1,0 +1,16 @@
+function showBasket() {
+  $("#basket").css("display","");
+  $("#searchForm").addClass("subject-selected");
+};
+
+var selectedLectureList = new Vue({
+  el: '#selectedLectureList',
+  data: {
+    lectureData: []
+  }
+});
+
+
+function updateSelectedLectureList(){
+  console.log(selectedLecture); 
+}
