--- frontend/webRoot/resources/js/mainVue.js
@@ -1,8 +1,0 @@
-var vueMajorList = new Vue({
-  el: '#majorListVue',
-  data: {
-    majorNames: ['sample',
-    'sample2',
-    'sample3']
-  }
-})
