+++ frontend/webRoot/resources/template/subjectTable.html
